feat: give these objects a name
It's much less friendly to use this library (and those that depend on
it, like npm-package-arg) when it's not clear whether the returned value
is a member of a class (which likely has some useful methods) or just an
object full of data that the user is expected to know what to do with.

This is especially vexing since some of this sort of data is saved to
various lock and config files, so _sometimes_ a thing that looks
suspiciously like a GitHost object actually is just a plain old JS
object.