fix: do not encodeURIComponent the domain
I ran into a situation where I have to test how Pacote handles git
shorthands from hosted-git-info, but I didn't want to ping GitHub to run
my tests.

I set up a git daemon and http server on localhost easily enough, but it
was still necessary to make it a 'real' HostedGit for the purposes of
npm-package-arg, and the : with port number was getting URL encoded.

This shows an example of doing this hack/test workaround, and fixes the
encoding issue.

It would be interesting to perhaps support GitHub Enterprise or on-prem
GitLab by letting users specify a prefix and format for hosted git info
in their npm configs, and then clone the definition with a different
domain name.