feat!: refactor
This also adds a `browseFile` method for browsing directly to a blob as
separate from a git tree. This is useful for hosts that serve blobs and
trees differently.

BREAKING CHANGE: `GitHost` now has a static `addHost` method to use
instead of manually editing the object from `lib/git-host-info.js`.